detector_placas/requirements.txt
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcrypt==3.2.2
blinker==1.7.0
certifi==2023.11.17
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.6
click==8.1.6
colorama==0.4.6
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
distro==1.9.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
pillow==10.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==3.0
pycurl==7.45.3
Pygments==2.17.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
service-identity==24.1.0
setuptools==68.1.2
six==1.16.0
ssh-import-id==5.11
Twisted==24.3.0
typing_extensions==4.10.0
urllib3==2.0.7
wadllib==1.3.6
wheel==0.42.0
zope.interface==6.1
